api/pyproject.toml
[tool.poetry]
name = "strawberryfields-api"
version = "0.1.0"
description = ""
authors = ["Ahn Shinho <fracsinus@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.111.1"}
argon2-cffi = "^23.1.0"
pydantic = "^2.8.2"
asyncpg = "^0.29.0"
boto3 = {extras = ["s3"], version = "^1.34.147"}


[tool.poetry.group.dev.dependencies]
typing-extensions = "^4.12.2"
black = "^24.4.2"
boto3-stubs = {extras = ["s3"], version = "^1.34.147"}

[tool.pyright]
reportMatchNotExhaustive = "error"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
